refactor(editor): extract the TipTap extension list into createEditorExtensions
Registers TaskLink, shares the Link mark's HTMLAttributes with pasted
pills, drops StarterKit's duplicate link/underline copies and renames
the mentionProjectId prop to projectId since it now also provides the
host project for task link pills.
