+++ frontend/src/components/tasks/partials/Comments.vue
@@ -113,7 +113,7 @@
 						:show-save="true"
 						:enable-discard-shortcut="true"
 						:enable-mentions="true"
-						:mention-project-id="projectId"
+						:project-id="projectId"
 						initial-mode="preview"
 						@update:modelValue="
 							() => {
@@ -184,7 +184,7 @@
 								:upload-callback="attachmentUpload"
 								:placeholder="$t('task.comment.placeholder')"
 								:enable-mentions="true"
-								:mention-project-id="projectId"
+								:project-id="projectId"
 								:storage-key="commentStorageKey"
 								@save="addComment()"
 							/>
